Combine TimetableSchedules and TimetableSchedulesMinutes to re-use CSS file
