+++ frontend/components/common/TimetableSchedules/index.tsx
@@ -1,15 +1,15 @@
 /* * */
 
 import type { Timetable } from '@/types/timetables.types';
+import type { Minute } from '@/types/timetables.types';
 
-import TimetableSchedulesMinute from '@/components/common/TimetableSchedulesMinute';
 import { useTranslations } from 'next-intl';
 
 import styles from './styles.module.css';
 
 /* * */
 
-interface Props {
+interface TimetableSchedulesProps {
 	selectedExceptionIds: string[]
 	setSelectedExceptionIds: (values: string[]) => void
 	timetableData: Timetable
@@ -17,7 +17,7 @@ interface Props {
 
 /* * */
 
-export default function Component({ selectedExceptionIds, setSelectedExceptionIds, timetableData }: Props) {
+export default function TimetableSchedules({ selectedExceptionIds, setSelectedExceptionIds, timetableData }: TimetableSchedulesProps) {
 	//
 
 	//
@@ -48,6 +48,57 @@ export default function Component({ selectedExceptionIds, setSelectedExceptionId
 				</div>
 			))}
 		</div>
+	);
+
+	//
+}
+
+/* * */
+
+interface TimetableSchedulesMinuteProps {
+	minuteData: Minute
+	selectedExceptionIds: string[]
+	setSelectedExceptionIds: (values: string[]) => void
+}
+
+/* * */
+
+function TimetableSchedulesMinute({ minuteData, selectedExceptionIds, setSelectedExceptionIds }: TimetableSchedulesMinuteProps) {
+	//
+
+	//
+	// A. Transform data
+
+	const isSelected = selectedExceptionIds.some(exceptionId => minuteData.exception_ids.includes(exceptionId));
+
+	//
+	// B. Handle actions
+
+	const handleMouseOverException = () => {
+		setSelectedExceptionIds(minuteData.exception_ids);
+	};
+
+	const handleMouseOutException = () => {
+		setSelectedExceptionIds([]);
+	};
+
+	//
+	// C. Render components
+
+	return (
+		<p
+			key={minuteData.minute_value}
+			className={`${styles.minute} ${minuteData.exception_ids.length > 0 && styles.withException} ${isSelected && styles.isSelected} ${!isSelected && selectedExceptionIds.length > 0 && styles.isOthersSelected}`}
+			onMouseOut={handleMouseOutException}
+			onMouseOver={handleMouseOverException}
+		>
+			{minuteData.minute_label}
+			{minuteData.exception_ids.length > 0 && minuteData.exception_ids.map(exceptionId => (
+				<span key={exceptionId} className={styles.exception}>
+					{exceptionId}
+				</span>
+			))}
+		</p>
 	);
 
 	//
