+++ frontend/components/common/TimetableSchedules/styles.module.css
@@ -80,8 +80,25 @@
 /* * */
 /* MINUTES */
 
+
 .minute {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	font-size: 14px;
+	line-height: 1;
+	padding: 0 6px;
 	font-weight: var(--font-weight-semibold);
+}
+
+.column:first-child .minute {
+	justify-content: flex-start;
+	padding-left: var(--size-spacing-15);
+	padding-right: var(--size-spacing-10);
+}
+
+.column:last-child .minute {
+	padding-right: var(--size-spacing-15);
 }
 
 .minute.withException {
@@ -90,6 +107,13 @@
 	font-weight: var(--font-weight-medium);
 }
 
+.minute.withException.isSelected {
+	color: var(--color-status-info-text);
+}
+
+.minute.isOthersSelected {
+	color: var(--color-system-text-400);
+}
 
 /* * */
 /* EXCEPTION */
@@ -97,7 +121,7 @@
 .exception {
 	position: absolute;
 	top: 0;
-	right: -1px;
 	font-size: 10px;
 	font-weight: var(--font-weight-bold);
+	transform: translate(12px, 0);
 }
