--- frontend/components/common/TimetableSchedulesMinute/styles.module.css
@@ -1,47 +1,0 @@
-/* * */
-/* CONTAINER */
-
-.container {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	font-size: 14px;
-	line-height: 1;
-	padding: 0 6px;
-	font-weight: var(--font-weight-semibold);
-}
-
-.column:first-child .container {
-	justify-content: flex-start;
-	padding-left: var(--size-spacing-15);
-	padding-right: var(--size-spacing-10);
-}
-
-.column:last-child .container {
-	padding-right: var(--size-spacing-15);
-}
-
-.container.withException {
-	position: relative;
-	color: var(--color-system-text-300);
-	font-weight: var(--font-weight-medium);
-}
-
-.container.withException.isSelected {
-	color: var(--color-status-info-text);
-}
-
-.container.isOthersSelected {
-	color: var(--color-system-text-400);
-}
-
-/* * */
-/* EXCEPTION */
-
-.exception {
-	position: absolute;
-	top: 0;
-	right: -1px;
-	font-size: 10px;
-	font-weight: var(--font-weight-bold);
-}
