+++ frontend/components/lines/PathStop/styles.module.css
@@ -10,7 +10,7 @@
 }
 
 .container:hover {
-	background-color: var(--color-system-background-200);
+	background-color: var(--color-system-background-300);
 }
 
 .container.isFirstStop {
@@ -36,7 +36,8 @@
 	display: flex;
 	flex-direction: column;
 	gap: var(--size-spacing-15);
-	padding: var(--size-spacing-20) 0;
+	padding-top: var(--size-spacing-20);
+	padding-bottom: var(--size-spacing-15);
 }
 
 .isFirstStop .detailsWrapper {
