+++ frontend/package.json
@@ -29,10 +29,10 @@
 		"inflected": "2.1.0",
 		"jsonwebtoken": "9.0.2",
 		"luxon": "3.5.0",
-		"maplibre-gl": "4.6.0",
+		"maplibre-gl": "4.7.0",
 		"minisearch": "7.1.0",
 		"next": "14.2.8",
-		"next-intl": "3.19.0",
+		"next-intl": "3.19.1",
 		"react": "18.3.1",
 		"react-dom": "18.3.1",
 		"react-map-gl": "7.1.7",
